+++ CameraPlus/Behaviours/CameraPreviewQuad.cs
@@ -66,19 +66,29 @@ namespace CameraPlus.Behaviours
 		public void SetCameraQuadPosition(string quadPosition, float cubeScale = 1)
         {
 			_cameraQuad.transform.localEulerAngles = new Vector3(0, 180, 0);
-			_cameraQuad.transform.localScale = new Vector3(0.15f * (PluginConfig.Instance.CameraQuadStretch ? 1.7778f : cam._cam.aspect), 0.15f * 1, 1); //1.7778f = 16 / 9
+			SetCameraQuadSize(cam.Config.cameraExtensions.previewCameraQuadScale, cam.Config.cameraExtensions.previewCameraMirrorMode);
 			if (quadPosition == "Top")
 				_cameraQuad.transform.localPosition = new Vector3(0, _cameraQuad.transform.localScale.y / 2 + _cameraCube.transform.localScale.y / 2, 0.05f * cubeScale);
 			else if (quadPosition == "Bottom")
 				_cameraQuad.transform.localPosition = new Vector3(0, -_cameraQuad.transform.localScale.y / 2 - _cameraCube.transform.localScale.y / 2, 0.05f * cubeScale);
 			else if (quadPosition == "Left")
-				_cameraQuad.transform.localPosition = new Vector3(_cameraQuad.transform.localScale.x / 2 + _cameraCube.transform.localScale.x / 2, 0, 0.05f * cubeScale);
+				_cameraQuad.transform.localPosition = new Vector3(Mathf.Abs(_cameraQuad.transform.localScale.x) / 2 + _cameraCube.transform.localScale.x / 2, 0, 0.05f * cubeScale);
+			else if (quadPosition == "Center")
+				_cameraQuad.transform.localPosition = new Vector3(0, 0, 0.15f * cubeScale);
 			else
-				_cameraQuad.transform.localPosition = new Vector3(-_cameraQuad.transform.localScale.x / 2 - _cameraCube.transform.localScale.x / 2, 0, 0.05f * cubeScale);
+				_cameraQuad.transform.localPosition = new Vector3(-Mathf.Abs(_cameraQuad.transform.localScale.x) / 2 - _cameraCube.transform.localScale.x / 2, 0, 0.05f * cubeScale);
 		}
 		public void SetCameraCubeSize(float cubeScale)
         {
 			_cameraCube.transform.localScale = new Vector3(0.15f * cubeScale, 0.15f * cubeScale, 0.22f * cubeScale);
+		}
+
+		public void SetCameraQuadSize(float quadScale, bool isMirror=false)
+        {
+			if(isMirror)
+				_cameraQuad.transform.localScale = new Vector3(-1 * 0.15f * (PluginConfig.Instance.CameraQuadStretch ? 1.7778f : cam._cam.aspect) * quadScale, 0.15f * 1 * quadScale, 1); //1.7778f = 16 / 9
+			else
+				_cameraQuad.transform.localScale = new Vector3(0.15f * (PluginConfig.Instance.CameraQuadStretch ? 1.7778f : cam._cam.aspect) * quadScale, 0.15f * 1 * quadScale, 1); //1.7778f = 16 / 9
 		}
 	}
 }