+++ CameraPlus/CHANGELOG.md
@@ -1,3 +1,8 @@
+# v6.7.1 Changes
+- The in-game camera preview is now in front of the white cube.  
+- It is now possible to resize the in-game camera preview.  
+- The in-game camera preview display can now be mirror-displayed.  
+
 # v6.7.0 Changes
 - Added the process to stop Camera Plus at the time of official Beatmap Editor.  
 - Change the referenced VMC Vatar to 1.0.0 or higher.